requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.4
email-validator==2.0.0
pytest==7.4.2
pytest-flask==1.2.0
uvicorn==0.29.0
fastapi==0.110.0
python-multipart==0.0.9
pydantic==1.10.13
langchain-mistralai
langchain-chroma
langchain-community
langchain-core
langchain-text-splitters==0.0.1
langsmith
python-jose
requests==2.31.0
SQLAlchemy==2.0.23
werkzeug==2.3.7
langchain
lark
watchfiles==0.20.0
python-jose
sqlalchemy-utils 